Create list characters

+++ WarcraftConsoleVersion/Game.cs
@@ -1,13 +1,25 @@
 ﻿public class Game
 {
     private string name;
-
+    public static List<Unit> character = new List<Unit>();
     public Game(string name)
     {
         this.name = name;
     }
 
-    internal static Unit AddUnit(string nameUnit, Unit unit)
+    public static int amountCharacter()
+    {
+        return character.Count();
+    }
+
+    public static void getAllCharacters()
+    {
+        for(int i = 0; i < character.Count; i++)
+        {
+            Console.WriteLine((1+i) + "." + character[i] + ", name: " + character[i].name);
+        }
+    }
+    public static void AddUnit(string nameUnit)
     {
         string codeUnit;
         bool createCheck = false;
@@ -18,27 +30,33 @@
             switch (codeUnit)
             {
                 case "1":
-                    unit = new GuardTowers(600, 120, nameUnit, 1, 800, 90, 3);
+                    GuardTowers guardTowers = new(600, 120, nameUnit, 1, 800, 90, 3);
+                    character.Add(guardTowers);
                     createCheck = true;
                     break;
                 case "2":
-                    unit = new Peasent(220, 75, nameUnit, 1, 99999);
+                    Peasent peasent = new(220, 75, nameUnit, 1, 99999);
+                    character.Add(peasent);
                     createCheck = true;
                     break;
                 case "3":
-                    unit = new Footman(420, 135, nameUnit, 1, 1, 13, 1, 2);
+                    Footman footman = new (420, 135, nameUnit, 1, 1, 13, 1, 2);
+                    character.Add(footman);
                     createCheck = true;
                     break;
                 case "4":
-                    unit = new Mage(550, 425, nameUnit, 1, 1, 27, 2, 2, 60, 285);
+                    Mage mage = new(550, 425, nameUnit, 1, 1, 27, 2, 2, 60, 285);
+                    character.Add(mage);
                     createCheck = true;
                     break;
                 case "5":
-                    unit = new Dragon(2200, 745, nameUnit, 1, 1, 64, 2, 6, 300, 600);
+                    Dragon dragon = new (2200, 745, nameUnit, 1, 1, 64, 2, 6, 300, 600);
+                    character.Add(dragon);
                     createCheck = true;
                     break;
                 case "6":
-                    unit = new Archer(550, 425, nameUnit, 1, 1, 27, 2, 2, 60);
+                    Archer archer= new(550, 425, nameUnit, 1, 1, 27, 2, 2, 60);
+                    character.Add(archer);
                     createCheck = true;
                     break;
             }
@@ -48,42 +66,60 @@
                 Console.WriteLine("Unknown code...");
             }
         }
-        return unit;
     }
 
-    public void GameStart(Unit oneCharachter, Unit twoCharachter)
+    public void GameStart(int one, int two)
     {
-        Thread myThread = new Thread(Fight);
-
-        myThread.Start();
-
-        while (oneCharachter.alive && twoCharachter.alive)
+        if(one != two)
         {
-            Console.WriteLine(twoCharachter.name + " attack " + oneCharachter.name +
-                ", left " + oneCharachter.healthPoint);
-            twoCharachter.Attack(oneCharachter);
-            Thread.Sleep(twoCharachter.getAttackSpeed() * 1000);
+            Unit oneCharachter = character[one];
+            Unit twoCharachter = character[two];
+            Thread myThread = new Thread(Fight);
 
-        }
+            myThread.Start();
 
-        void Fight()
-        {
             while (oneCharachter.alive && twoCharachter.alive)
             {
-                Console.WriteLine(oneCharachter.name + " attack " + twoCharachter.name +
-                    ", left " + twoCharachter.healthPoint);
-                oneCharachter.Attack(twoCharachter);
-                Thread.Sleep(oneCharachter.getAttackSpeed() * 1000);
+                Console.WriteLine(twoCharachter.name + " attack " + oneCharachter.name +
+                    ", left " + oneCharachter.healthPoint);
+                twoCharachter.Attack(oneCharachter);
+                Thread.Sleep(twoCharachter.getAttackSpeed() * 1000);
+
+            }
+
+            void Fight()
+            {
+                while (oneCharachter.alive && twoCharachter.alive)
+                {
+                    Console.WriteLine(oneCharachter.name + " attack " + twoCharachter.name +
+                        ", left " + twoCharachter.healthPoint);
+                    oneCharachter.Attack(twoCharachter);
+                    Thread.Sleep(oneCharachter.getAttackSpeed() * 1000);
+                }
+            }
+
+            if (oneCharachter.alive)
+            {
+                Console.WriteLine(oneCharachter.name + " kill " + twoCharachter.name);
+            }
+            else
+            {
+                Console.WriteLine(twoCharachter.name + " kill " + oneCharachter.name);
             }
         }
+    }
 
-        if (oneCharachter.alive)
+    public static void AddCharaters()
+    {
+        string nameUnit = "none";
+        while (nameUnit != "")
         {
-            Console.WriteLine(oneCharachter.name + " kill " + twoCharachter.name);
-        }
-        else
-        {
-            Console.WriteLine(twoCharachter.name + " kill " + oneCharachter.name);
+            Console.WriteLine("\nWrite name " + (Game.amountCharacter() + 1) + " character:");
+            nameUnit = Console.ReadLine();
+            if (nameUnit != "")
+            {
+                Game.AddUnit(nameUnit);
+            }
         }
     }
 }
